// src/services/model/evaLogins.ts
/**
 * PostgREST API
 * standard public schema
 *
 * OpenAPI spec version: 7.0.0 (2b61a63)
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface EvaLogins { 
    /**
     * Note: This is a Primary Key.<pk/>
     */
    idlogin: number;
    idusuario?: string;
    /**
     * Note: This is a Foreign Key to `eva_examenes.idexamen`.<fk table='eva_examenes' column='idexamen'/>
     */
    idexamen: number;
    /**
     * Note: This is a Foreign Key to `eva_admin_usuarios.idusuario`.<fk table='eva_admin_usuarios' column='idusuario'/>
     */
    idadmin?: string;
    token?: string;
    macadress?: string;
    ipautorizacion?: string;
    horaIni?: string;
    horaFin?: string;
    fecharegistro?: string;
}