// src/services/model/evaUlasalle.ts
/**
 * Desarrollo Albertano del Hoyo!
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * Do not edit the class manually.
 */

export interface EvaUlasalle { 
    /**
     * Note: This is a Primary Key.<pk/>
     */
    estado?: string;
    grupos?: string;
    telefonoCE?: string;
    fechaNacimiento?: string;
    nombres?: string;
    dni?: string;
    apellidos?: string;
    email?: string;    
    token?:any;
}


export interface EvaUlasalleLogin { 
    /**
     * Note: This is a Primary Key.<pk/>
     */
    dni: string;
    tokenid: string;
    navegador: string;
    ip: string;
    ubicacion: string;

}